--- templates/.mason/bricks.json
@@ -1,1 +1,0 @@
-{"simple_rest_api_dartfrog":"/Users/codekeyz/Projects/Invertase.io/globe/templates/simple_rest_api_dartfrog"}
