--- templates/hello_world/.dart_frog/server.dart
@@ -1,46 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint, implicit_dynamic_list_literal
-
-import 'dart:io';
-
-import 'package:dart_frog/dart_frog.dart';
-
-
-import '../routes/index.dart' as index;
-import '../routes/api/[id].dart' as api_$id;
-
-import '../routes/_middleware.dart' as middleware;
-
-void main() async {
-  final address = InternetAddress.tryParse('') ?? InternetAddress.anyIPv6;
-  final port = int.tryParse(Platform.environment['PORT'] ?? '3000') ?? 3000;
-  hotReload(() => createServer(address, port));
-}
-
-Future<HttpServer> createServer(InternetAddress address, int port) {
-  final handler = Cascade().add(buildRootHandler()).handler;
-  return serve(handler, address, port);
-}
-
-Handler buildRootHandler() {
-  final pipeline = const Pipeline().addMiddleware(middleware.middleware);
-  final router = Router()
-    ..mount('/api', (context) => buildApiHandler()(context))
-    ..mount('/', (context) => buildHandler()(context));
-  return pipeline.addHandler(router);
-}
-
-Handler buildApiHandler() {
-  final pipeline = const Pipeline();
-  final router = Router()
-    ..all('/<id>', (context,id,) => api_$id.onRequest(context,id,));
-  return pipeline.addHandler(router);
-}
-
-Handler buildHandler() {
-  final pipeline = const Pipeline();
-  final router = Router()
-    ..all('/', (context) => index.onRequest(context,));
-  return pipeline.addHandler(router);
-}
-
