--- templates/hello_world/README.md
@@ -1,3 +1,0 @@
-# hello_world
-
-
