--- templates/hello_world/pubspec.yaml
@@ -1,13 +1,0 @@
-name: hello_world
-description: A simple dart frog project
-version: 0.1.0+1
-environment:
-  sdk: '>=3.0.0'
-
-dependencies:
-  dart_frog: ">=1.0.0 <2.0.0"
-  intl: ^0.19.0
-
-dev_dependencies:
-  test: ^1.25.4
-  mocktail: ^1.0.3
