--- templates/hello_world/routes/_middleware.dart
@@ -1,15 +1,0 @@
-import 'package:dart_frog/dart_frog.dart';
-
-Handler middleware(Handler handler) {
-  return (context) async {
-    // Execute code before request is handled.
-
-    // Forward the request to the respective handler.
-    final response = await handler(context);
-
-    // Execute code after request is handled.
-
-    // Return a response.
-    return response;
-  };
-}
