--- templates/mason-lock.json
@@ -1,1 +1,0 @@
-{"bricks":{"simple_rest_api_dartfrog":{"path":"/Users/codekeyz/Projects/Invertase.io/globe/templates/simple_rest_api_dartfrog"}}}
