+++ templates/simple_rest_api_dartfrog/.gitignore
@@ -1,0 +1,12 @@
+.DS_Store
+node_modules/
+npm-debug.log
+dist/
+.vscode/
+.wrangler
+.nvim.lua
+.idea
+*.iml
+.dart_tool
+.fvm
+.dart_frog
