--- templates/simple_rest_api_dartfrog/__brick__/{{project_name.snakeCase()}}/README.md
@@ -1,3 +1,0 @@
-# {{project_name}}
-
-{{package_description}}
