--- templates/simple_rest_api_dartfrog/__brick__/{{project_name.snakeCase()}}/lib/routes/_middleware.dart
@@ -1,1 +1,0 @@
-// Global middleware can be defined here
