--- templates/simple_rest_api_dartfrog/__brick__/{{project_name.snakeCase()}}/lib/routes/index.dart
@@ -1,11 +1,0 @@
-import 'package:dart_frog/dart_frog.dart';
-import 'package:intl/intl.dart';
-
-Response onRequest(RequestContext context) {
-  final String formattedDate = DateFormat('yyyy-MM-dd – kk:mm').format(
-    DateTime.now(),
-  );
-  return Response(
-    body: 'Welcome to {{project_name}} REST API! Current time: $formattedDate',
-  );
-}
