--- templates/simple_rest_api_dartfrog/__brick__/{{project_name.snakeCase()}}/pubspec.yaml
@@ -1,13 +1,0 @@
-name: {{project_name.snakeCase()}}
-description: A simple dart frog project
-version: 0.1.0+1
-environment:
-  sdk: '>=3.0.0'
-
-dependencies:
-  dart_frog: ^0.1.0
-  intl: ^0.19.0
-
-dev_dependencies:
-  test: ^1.25.4
-  mocktail: ^1.0.3
