--- templates/simple_rest_api_dartfrog/brick.yaml
@@ -1,11 +1,0 @@
-name: simple_rest_api_dartfrog
-description: A highly customizable template for creating RESTful servers with Dart.
-version: 0.1.0+1
-environment:
-  mason: ">=0.1.0-dev.52 <0.1.0"
-vars:
-  project_name:
-    type: string
-    description: Set a custom name for your project
-    default: simple_rest_api_dartfrog
-    prompt: Provide name for project?
