--- templates/simple_rest_api_dartfrog/hooks/.gitignore
@@ -1,1 +1,0 @@
-.dart_tool
