--- templates/simple_rest_api_dartfrog/hooks/pubspec.yaml
@@ -1,7 +1,0 @@
-name: mason_hooks
-
-dependencies:
-  mason:
-
-environment:
-  sdk: "^3.2.0"
