+++ templates/simple_rest_api_dartfrog/routes/index.dart
@@ -1,0 +1,11 @@
+import 'package:dart_frog/dart_frog.dart';
+import 'package:intl/intl.dart';
+
+Response onRequest(RequestContext context) {
+  final String formattedDate = DateFormat('yyyy-MM-dd – kk:mm').format(
+    DateTime.now(),
+  );
+  return Response(
+    body: 'Welcome to DartFrog REST API! Current time: $formattedDate',
+  );
+}
